requirements.txt
asgiref==3.6.0
chardet==5.1.0
Django==4.2.1
django-csvimport==2.17
Pillow==9.5.0
reportlab==3.6.12
sqlparse==0.4.4
python-dotenv==1.2.1